requirements.txt
asgiref==3.7.2
Django==4.2.7
django-crispy-forms==2.1
django-modeltranslation==0.18.11
gunicorn==21.2.0
packaging==23.2
Paginator==0.5.1
Pillow==10.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
sorl-thumbnail==12.10.0
sqlparse==0.4.4
typing_extensions==4.8.0
